Play the "Suspense" music when Mister Dark kidnaps Betilla
Only tested on one version, but the pointer was really easy to find so
it should work equally well on all versions. If I had realized that, I
would have implemented it sooner! It became imperative when I realized
that without this intervention, the "MIDI" track for beating the boss
would still play over the kidnapping vignette, which would be very
incongruous...